Create electronicstest.c
This is a code that sends high, then medium and then low pwm signals to a gpio pin on the raspberry pi. The idea of this code is that you can run it first of all to test that the connected electronics are working and to see what they do in response to each signal, the "gpiopwm.c" program can then have the values in it changed to ensure the motors have the correct response.